requirements.txt
aniso8601==8.1.0
astroid==2.3.3
attrs==20.3.0
bitarray==1.6.1
blis==0.7.4
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
cmd2==1.4.0
cursor==1.3.4
cycler==0.10.0
cymem==2.0.5
dnspython==1.16.0
en-core-web-sm==2.3.1
eventlet==0.29.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
gunicorn==20.0.4
halo==0.0.28
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014535162/work
lazy-object-proxy==1.4.3
log-symbols==0.0.14
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603378225747/work
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
olefile==0.46
opencv-python==4.4.0.46
ordered-set==3.1.1
ordered-set-stubs==0.1.3
pandas @ file:///tmp/build/80754af9/pandas_1602088120436/work
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.5
profanity-check==1.0.3
profanity-filter==1.3.3
pycparser==2.20
pycryptodome==3.9.4
pydantic==1.7.3
pylint==2.7.4
pyparsing==2.4.7
pyperclip==1.7.0
python-dateutil==2.8.1
python-engineio==3.13.2
pytz==2020.1
redis==3.5.3
requests==2.25.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376899566/work
scipy==1.5.4
simpleaudio==1.0.4
sip==4.19.13
six==1.15.0
spacy==2.3.5
spinners==0.0.23
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.1
tornado==6.0.4
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.4
wasabi==0.8.0
wcwidth==0.2.5
Werkzeug==1.0.1
